+++ src/interface/serviceResponse/user/autoLoginToGetTenantTokenDTO.ts
