+++ src/interface/thingsboardConnector/user/TBCreateTenantAdminDTO.ts
