+++ src/interface/thingsboardConnector/user/TBCreateTenantDTO.ts
