+++ src/interface/thingsboardConnector/user/TBLoginSystemAdminDTO.ts
