+++ src/interface/thingsboardConnector/user/TBLoginTenantDTO.ts
