+++ src/interface/thingsboardConnector/user/TBSearchTenantAdminDTO.ts
