+++ src/interface/thingsboardConnector/user/TBUserInterface.ts
