+++ src/library/thingsboardConnecter/user/loginByTenantId.ts
@@ -1,7 +1,7 @@
 import { TB_SERVER } from '../../../constants/env';
 import APICaller from '../../../helpers/apiCaller';
 import WinstonLogger from '../../../helpers/loggers';
-import TBLoginTenantDTO from '../../../interface/thingsboardConnector/TBLoginTenantDTO';
+import TBLoginTenantDTO from '../../../interface/thingsboardConnector/user/TBLoginTenantDTO';
 
 const loggers = new WinstonLogger({ type: 'Tenant' });
