+++ src/router/TBUser.routes.ts
@@ -1,10 +1,20 @@
 import express from 'express';
+import createDevice from '../library/thingsboardConnecter/device/createDevice';
 import { autoLoginTenantAccount } from '../controller/user.controller';
+import { autoLoginToGetTenantToken } from '../service/user';
 
 const router = express.Router();
 
-router.get('/test', (req, res) => {
-    res.json('echo');
+const test = {
+    name: `string${Date.now()}`,
+    type: 'string',
+    label: 'string',
+};
+
+router.get('/test', async (req, res) => {
+    const tenant = await autoLoginToGetTenantToken(req.headers.TBAdminToken as string);
+    const r = await createDevice(tenant.token, test);
+    res.json(r);
 });
 
 // 若tenant或是tenant admin不存在會自動create
