+++ src/service/user/helpers/isEntityListEmpty.ts
@@ -1,5 +1,5 @@
-import TBSearchTenantAdminDTO from '../../../interface/thingsboardConnector/TBSearchTenantAdminDTO';
-import TBSearchTenantDTO from '../../../interface/thingsboardConnector/TBSearchTenantDTO';
+import TBSearchTenantAdminDTO from '../../../interface/thingsboardConnector/user/TBSearchTenantAdminDTO';
+import TBSearchTenantDTO from '../../../interface/thingsboardConnector/user/TBSearchTenantDTO';
 
 export default function isEntityListEmpty(tenantInfo: TBSearchTenantDTO | TBSearchTenantAdminDTO) {
     if (tenantInfo.data.length < 1) {
