+++ src/service/user/helpers/isTenantEntityNotFound.ts
@@ -1,7 +1,7 @@
 import checkStatusError from '../../../helpers/checkStatusError';
 import checkTenantAdminOrTenantEmpty from './isEntityListEmpty';
-import TBSearchTenantAdminDTO from '../../../interface/thingsboardConnector/TBSearchTenantAdminDTO';
-import TBSearchTenantDTO from '../../../interface/thingsboardConnector/TBSearchTenantDTO';
+import TBSearchTenantAdminDTO from '../../../interface/thingsboardConnector/user/TBSearchTenantAdminDTO';
+import TBSearchTenantDTO from '../../../interface/thingsboardConnector/user/TBSearchTenantDTO';
 
 export default function isTenantEntityNotFound(entity: TBSearchTenantDTO | TBSearchTenantAdminDTO) {
     if (checkStatusError(entity)) {
